Fix a bug and optimize IntrinsicsSourcesService
Fix a bug in IntrinsicsSourcesService where the data was not showing up when typing methods filter for the first time due to a misconception about how asynchronous code was used.

Optimize IntrinsicsSourcesService by adding asynchronous fetching of content from github and compiling of all content instead of its parts separately.